rename public to lib
