+++ template-component/src/component/lib.ts
